:art: reformat + :wrench: compile_commands.json

+++ src/complex/cimagl.c
@@ -1,6 +1,3 @@
 #include "complex_impl.h"
 
-long double (cimagl)(long double complex z)
-{
-	return cimagl(z);
-}
+long double(cimagl)(long double complex z) { return cimagl(z); }
